+++ plugins/main/public/components/overview/vulnerabilities/events/vulnerabilities-columns.tsx
@@ -1,27 +1,26 @@
 import { tDataGridColumn } from '../../../common/data-grid';
+import { commonColumns } from '../../common/data-grid-columns';
 
 export const vulnerabilitiesColumns: tDataGridColumn[] = [
-  {
-    id: 'timestamp',
-    isSortable: true,
-    defaultSortDirection: 'desc',
-  },
-  {
-    id: 'agent.name',
-  },
+  commonColumns.timestamp,
+  commonColumns['agent.name'],
   {
     id: 'data.vulnerability.cve',
+    initialWidth: 186,
   },
   {
     id: 'data.vulnerability.severity',
+    initialWidth: 217,
   },
   {
     id: 'data.vulnerability.package.name',
+    initialWidth: 257,
   },
   {
     id: 'data.vulnerability.package.version',
   },
   {
     id: 'data.vulnerability.status',
+    initialWidth: 199,
   },
 ];